chore: move request related pages into Requests folder
