+++ NextGen-BM-FE/src/App.tsx
@@ -8,9 +8,9 @@ import MainLayout from "./layouts/MainLayout";
 import LoginPage from "./pages/Account/LoginPage";
 import HomePage from "./pages/HomePage";
 import { FC } from "react";
-import CreateRequestPage from "./pages/CreateRequestPage";
-import RequestsListPage from "./pages/RequestsListPage";
-import RequestPage from "./pages/RequestPage";
+import CreateRequestPage from "./pages/Requests/CreateRequestPage";
+import RequestsListPage from "./pages/Requests/RequestsListPage";
+import RequestPage from "./pages/Requests/RequestPage";
 import BuildingExpensesPage from "./pages/Buildings/BuildingExpensesPage";
 import PropertyFeesPage from "./pages/Properties/PropertyFeesPage";
 import PaymentPage from "./pages/PaymentPage";
